Integrate create profile with import flow (#188510)
