+++ src/vs/workbench/contrib/userDataProfile/browser/userDataProfile.ts
@@ -3,18 +3,16 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import 'vs/css!./media/userDataProfileCreateWidget';
 import { Disposable, DisposableStore, IDisposable, MutableDisposable } from 'vs/base/common/lifecycle';
 import { isWeb } from 'vs/base/common/platform';
-import { Event } from 'vs/base/common/event';
 import { ServicesAccessor } from 'vs/editor/browser/editorExtensions';
 import { localize } from 'vs/nls';
 import { Action2, IMenuService, ISubmenuItem, MenuId, MenuRegistry, registerAction2 } from 'vs/platform/actions/common/actions';
 import { ContextKeyExpr, IContextKey, IContextKeyService } from 'vs/platform/contextkey/common/contextkey';
-import { IUserDataProfile, IUserDataProfilesService, ProfileResourceType, UseDefaultProfileFlags } from 'vs/platform/userDataProfile/common/userDataProfile';
+import { IUserDataProfile, IUserDataProfilesService } from 'vs/platform/userDataProfile/common/userDataProfile';
 import { IWorkbenchContribution } from 'vs/workbench/common/contributions';
 import { ILifecycleService, LifecyclePhase } from 'vs/workbench/services/lifecycle/common/lifecycle';
-import { CURRENT_PROFILE_CONTEXT, HAS_PROFILES_CONTEXT, IS_CURRENT_PROFILE_TRANSIENT_CONTEXT, IS_PROFILE_IMPORT_IN_PROGRESS_CONTEXT, IUserDataProfileImportExportService, IUserDataProfileManagementService, IUserDataProfileService, PROFILES_CATEGORY, PROFILE_FILTER, IS_PROFILE_EXPORT_IN_PROGRESS_CONTEXT, ProfilesMenu, PROFILES_ENABLEMENT_CONTEXT, PROFILES_TITLE } from 'vs/workbench/services/userDataProfile/common/userDataProfile';
+import { CURRENT_PROFILE_CONTEXT, HAS_PROFILES_CONTEXT, IS_CURRENT_PROFILE_TRANSIENT_CONTEXT, IS_PROFILE_IMPORT_IN_PROGRESS_CONTEXT, IUserDataProfileImportExportService, IUserDataProfileManagementService, IUserDataProfileService, PROFILES_CATEGORY, PROFILE_FILTER, IS_PROFILE_EXPORT_IN_PROGRESS_CONTEXT, ProfilesMenu, PROFILES_ENABLEMENT_CONTEXT, PROFILES_TITLE, IProfileTemplateInfo } from 'vs/workbench/services/userDataProfile/common/userDataProfile';
 import { IQuickInputService, IQuickPickItem, IQuickPickSeparator } from 'vs/platform/quickinput/common/quickInput';
 import { INotificationService } from 'vs/platform/notification/common/notification';
 import { IFileDialogService } from 'vs/platform/dialogs/common/dialogs';
@@ -25,22 +23,6 @@ import { IWorkspaceTagsService } from 'vs/workbench/contrib/tags/common/workspac
 import { getErrorMessage } from 'vs/base/common/errors';
 import { Categories } from 'vs/platform/action/common/actionCommonCategories';
 import { IOpenerService } from 'vs/platform/opener/common/opener';
-import { IProductService } from 'vs/platform/product/common/productService';
-import { IRequestService, asJson } from 'vs/platform/request/common/request';
-import { CancellationToken } from 'vs/base/common/cancellation';
-import { ILogService } from 'vs/platform/log/common/log';
-import Severity from 'vs/base/common/severity';
-import { $, append } from 'vs/base/browser/dom';
-import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
-import { ISelectOptionItem, SelectBox } from 'vs/base/browser/ui/selectBox/selectBox';
-import { IContextViewService } from 'vs/platform/contextview/browser/contextView';
-import { defaultSelectBoxStyles } from 'vs/platform/theme/browser/defaultStyles';
-import { isString } from 'vs/base/common/types';
-
-interface IProfileTemplateInfo {
-	readonly name: string;
-	readonly url: string;
-}
 
 type IProfileTemplateQuickPickItem = IQuickPickItem & IProfileTemplateInfo;
 
@@ -59,14 +41,7 @@ export class UserDataProfilesWorkbenchContribution extends Disposable implements
 		@IWorkspaceContextService private readonly workspaceContextService: IWorkspaceContextService,
 		@IWorkspaceTagsService private readonly workspaceTagsService: IWorkspaceTagsService,
 		@IContextKeyService contextKeyService: IContextKeyService,
-		@IQuickInputService private readonly quickInputService: IQuickInputService,
-		@INotificationService private readonly notificationService: INotificationService,
 		@ILifecycleService private readonly lifecycleService: ILifecycleService,
-		@IProductService private readonly productService: IProductService,
-		@IRequestService private readonly requestService: IRequestService,
-		@IInstantiationService private readonly instantiationService: IInstantiationService,
-		@IContextViewService private readonly contextViewService: IContextViewService,
-		@ILogService private readonly logService: ILogService,
 	) {
 		super();
 
@@ -228,7 +203,7 @@ export class UserDataProfilesWorkbenchContribution extends Disposable implements
 				});
 			}
 			run() {
-				return that.saveProfile(that.userDataProfileService.currentProfile);
+				return that.userDataProfileImportExportService.editProfile(that.userDataProfileService.currentProfile);
 			}
 		});
 	}
@@ -375,7 +350,7 @@ export class UserDataProfilesWorkbenchContribution extends Disposable implements
 					}
 					try {
 						if ((<IProfileTemplateQuickPickItem>selectedItem).url) {
-							return await that.saveProfile(undefined, (<IProfileTemplateQuickPickItem>selectedItem).url);
+							return await that.userDataProfileImportExportService.createProfile(URI.parse((<IProfileTemplateQuickPickItem>selectedItem).url));
 						}
 						const profile = selectedItem.label === quickPick.value ? URI.parse(quickPick.value) : await this.getProfileUriFromFileSystem(fileDialogService);
 						if (profile) {
@@ -433,192 +408,9 @@ export class UserDataProfilesWorkbenchContribution extends Disposable implements
 			}
 
 			run(accessor: ServicesAccessor) {
-				return that.saveProfile(undefined, that.userDataProfileService.currentProfile);
+				return that.userDataProfileImportExportService.createProfile(that.userDataProfileService.currentProfile);
 			}
 		}));
-	}
-
-	private async saveProfile(profile: IUserDataProfile): Promise<void>;
-	private async saveProfile(profile?: IUserDataProfile, source?: IUserDataProfile | string): Promise<void>;
-	private async saveProfile(profile?: IUserDataProfile, source?: IUserDataProfile | string): Promise<void> {
-
-		type CreateProfileInfoClassification = {
-			owner: 'sandy081';
-			comment: 'Report when profile is about to be created';
-		};
-		this.telemetryService.publicLog2<{}, CreateProfileInfoClassification>('userDataProfile.startCreate');
-
-		const disposables = new DisposableStore();
-		const title = profile ? localize('save profile', "Edit {0} Profile...", profile.name) : localize('create new profle', "Create New Profile...");
-
-		const settings: IQuickPickItem & { id: ProfileResourceType } = { id: ProfileResourceType.Settings, label: localize('settings', "Settings"), picked: !profile?.useDefaultFlags?.settings };
-		const keybindings: IQuickPickItem & { id: ProfileResourceType } = { id: ProfileResourceType.Keybindings, label: localize('keybindings', "Keyboard Shortcuts"), picked: !profile?.useDefaultFlags?.keybindings };
-		const snippets: IQuickPickItem & { id: ProfileResourceType } = { id: ProfileResourceType.Snippets, label: localize('snippets', "User Snippets"), picked: !profile?.useDefaultFlags?.snippets };
-		const tasks: IQuickPickItem & { id: ProfileResourceType } = { id: ProfileResourceType.Tasks, label: localize('tasks', "User Tasks"), picked: !profile?.useDefaultFlags?.tasks };
-		const extensions: IQuickPickItem & { id: ProfileResourceType } = { id: ProfileResourceType.Extensions, label: localize('extensions', "Extensions"), picked: !profile?.useDefaultFlags?.extensions };
-		const resources = [settings, keybindings, snippets, tasks, extensions];
-
-		const quickPick = this.quickInputService.createQuickPick();
-		quickPick.title = title;
-		quickPick.placeholder = localize('name placeholder', "Profile name");
-		quickPick.value = profile?.name ?? '';
-		quickPick.canSelectMany = true;
-		quickPick.matchOnDescription = false;
-		quickPick.matchOnDetail = false;
-		quickPick.matchOnLabel = false;
-		quickPick.sortByLabel = false;
-		quickPick.hideCountBadge = true;
-		quickPick.ok = false;
-		quickPick.customButton = true;
-		quickPick.hideCheckAll = true;
-		quickPick.ignoreFocusOut = true;
-		quickPick.customLabel = profile ? localize('save', "Save") : localize('create', "Create");
-		quickPick.description = localize('customise the profile', "Choose what to configure in your Profile:");
-		quickPick.items = [...resources];
-
-		const update = () => {
-			quickPick.items = resources;
-			quickPick.selectedItems = resources.filter(item => item.picked);
-		};
-		update();
-
-		const validate = () => {
-			if (!profile && this.userDataProfilesService.profiles.some(p => p.name === quickPick.value)) {
-				quickPick.validationMessage = localize('profileExists', "Profile with name {0} already exists.", quickPick.value);
-				quickPick.severity = Severity.Warning;
-				return;
-			}
-			if (resources.every(resource => !resource.picked)) {
-				quickPick.validationMessage = localize('invalid configurations', "The profile should contain at least one configuration.");
-				quickPick.severity = Severity.Warning;
-				return;
-			}
-			quickPick.severity = Severity.Ignore;
-			quickPick.validationMessage = undefined;
-		};
-
-		disposables.add(quickPick.onDidChangeSelection(items => {
-			let needUpdate = false;
-			for (const resource of resources) {
-				resource.picked = items.includes(resource);
-				const description = resource.picked ? undefined : localize('use default profile', "Using Default Profile");
-				if (resource.description !== description) {
-					resource.description = description;
-					needUpdate = true;
-				}
-			}
-			if (needUpdate) {
-				update();
-			}
-			validate();
-		}));
-
-		disposables.add(quickPick.onDidChangeValue(validate));
-
-		let result: { name: string; items: ReadonlyArray<IQuickPickItem> } | undefined;
-		disposables.add(Event.any(quickPick.onDidCustom, quickPick.onDidAccept)(() => {
-			if (!quickPick.value) {
-				quickPick.validationMessage = localize('name required', "Provide a name for the new profile");
-				quickPick.severity = Severity.Error;
-			}
-			if (quickPick.validationMessage) {
-				return;
-			}
-			result = { name: quickPick.value, items: quickPick.selectedItems };
-			quickPick.hide();
-			quickPick.severity = Severity.Ignore;
-			quickPick.validationMessage = undefined;
-		}));
-
-		if (!profile) {
-			const domNode = $('.profile-type-widget');
-			append(domNode, $('.profile-type-create-label', undefined, localize('create from', "Copy from:")));
-			const separator = { text: '\u2500\u2500\u2500\u2500\u2500\u2500\u2500\u2500\u2500\u2500\u2500', isDisabled: true };
-			const profileOptions: (ISelectOptionItem & { id?: string; source?: IUserDataProfile | string })[] = [];
-			profileOptions.push({ text: localize('empty profile', "None") });
-			const templates = await this.getProfileTemplatesFromProduct();
-			if (templates.length) {
-				profileOptions.push({ ...separator, decoratorRight: localize('from templates', "Profile Templates") });
-				for (const template of templates) {
-					profileOptions.push({ text: template.name, id: template.url, source: template.url });
-				}
-			}
-			profileOptions.push({ ...separator, decoratorRight: localize('from existing profiles', "Existing Profiles") });
-			for (const profile of this.userDataProfilesService.profiles) {
-				profileOptions.push({ text: profile.name, id: profile.id, source: profile });
-			}
-
-			const findOptionIndex = () => {
-				const index = profileOptions.findIndex(option => {
-					if (isString(source)) {
-						return option.id === source;
-					} else if (source) {
-						return option.id === source.id;
-					}
-					return false;
-				});
-				return index > -1 ? index : 0;
-			};
-
-			const selectBox = disposables.add(this.instantiationService.createInstance(SelectBox, profileOptions, findOptionIndex(), this.contextViewService, defaultSelectBoxStyles, { useCustomDrawn: true }));
-			selectBox.render(append(domNode, $('.profile-type-select-container')));
-			quickPick.widget = domNode;
-
-			const updateQuickpickInfo = () => {
-				const option = profileOptions[findOptionIndex()];
-				for (const resource of resources) {
-					resource.picked = option.source && !isString(option.source) ? !option.source?.useDefaultFlags?.[resource.id] : true;
-				}
-				update();
-			};
-
-			updateQuickpickInfo();
-			disposables.add(selectBox.onDidSelect(({ index }) => {
-				source = profileOptions[index].source;
-				updateQuickpickInfo();
-			}));
-		}
-
-		quickPick.show();
-
-		await new Promise<void>((c, e) => {
-			disposables.add(quickPick.onDidHide(() => {
-				disposables.dispose();
-				c();
-			}));
-		});
-
-		if (!result) {
-			this.telemetryService.publicLog2<{}, CreateProfileInfoClassification>('userDataProfile.cancelCreate');
-			return;
-		}
-
-		this.telemetryService.publicLog2<{}, CreateProfileInfoClassification>('userDataProfile.successCreate');
-
-		try {
-			const useDefaultFlags: UseDefaultProfileFlags | undefined = result.items.length === resources.length
-				? undefined
-				: {
-					settings: !result.items.includes(settings),
-					keybindings: !result.items.includes(keybindings),
-					snippets: !result.items.includes(snippets),
-					tasks: !result.items.includes(tasks),
-					extensions: !result.items.includes(extensions)
-				};
-			if (profile) {
-				await this.userDataProfileManagementService.updateProfile(profile, { name: result.name, useDefaultFlags: profile.useDefaultFlags && !useDefaultFlags ? {} : useDefaultFlags });
-			} else {
-				if (isString(source)) {
-					await this.userDataProfileImportExportService.importProfile(URI.parse(source), { mode: 'apply', name: result.name, useDefaultFlags });
-				} else if (source) {
-					await this.userDataProfileImportExportService.createFromProfile(source, result.name, { useDefaultFlags });
-				} else {
-					await this.userDataProfileManagementService.createAndEnterProfile(result.name, { useDefaultFlags });
-				}
-			}
-		} catch (error) {
-			this.notificationService.error(error);
-		}
 	}
 
 	private registerCreateProfileAction(): void {
@@ -646,7 +438,7 @@ export class UserDataProfilesWorkbenchContribution extends Disposable implements
 			}
 
 			async run(accessor: ServicesAccessor) {
-				return that.saveProfile();
+				return that.userDataProfileImportExportService.createProfile();
 			}
 		}));
 	}
@@ -726,7 +518,7 @@ export class UserDataProfilesWorkbenchContribution extends Disposable implements
 
 	private async getProfileTemplatesQuickPickItems(): Promise<IProfileTemplateQuickPickItem[]> {
 		const quickPickItems: IProfileTemplateQuickPickItem[] = [];
-		const profileTemplates = await this.getProfileTemplatesFromProduct();
+		const profileTemplates = await this.userDataProfileManagementService.getBuiltinProfileTemplates();
 		for (const template of profileTemplates) {
 			quickPickItems.push({
 				label: template.name,
@@ -734,22 +526,6 @@ export class UserDataProfilesWorkbenchContribution extends Disposable implements
 			});
 		}
 		return quickPickItems;
-	}
-
-	private async getProfileTemplatesFromProduct(): Promise<IProfileTemplateInfo[]> {
-		if (this.productService.profileTemplatesUrl) {
-			try {
-				const context = await this.requestService.request({ type: 'GET', url: this.productService.profileTemplatesUrl }, CancellationToken.None);
-				if (context.res.statusCode === 200) {
-					return (await asJson<IProfileTemplateInfo[]>(context)) || [];
-				} else {
-					this.logService.error('Could not get profile templates.', context.res.statusCode);
-				}
-			} catch (error) {
-				this.logService.error(error);
-			}
-		}
-		return [];
 	}
 
 	private async reportWorkspaceProfileInfo(): Promise<void> {
